Remove the `TableReaddirExt` trait
